+++ InuyashaRescueTeam/InuyashaRescueTeam/InuyashaRescueTeam.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="Source Files">
@@ -98,6 +98,7 @@
     </ClCompile>
     <ClCompile Include="src\InputManager\InputManager.cpp">
       <Filter>Source Files\InputManager</Filter>
+    </ClCompile>
     <ClCompile Include="src\Map\MapManager.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -150,6 +151,7 @@
     </ClInclude>
     <ClInclude Include="include\InputManager\InputManager.h">
       <Filter>Header Files\InputManager</Filter>
+    </ClInclude>
     <ClInclude Include="include\Map\MapManager.h">
       <Filter>Header Files</Filter>
     </ClInclude>
