+++ InuyashaRescueTeam/InuyashaRescueTeam/src/InputManager/InputManager.cpp
@@ -1,4 +1,4 @@
-#include "InputManager.h"
+#include "InputManager/InputManager.h"
 
 // static 멤버 변수 초기화
 InputManager& InputManager::GetInstance()
@@ -7,30 +7,68 @@ InputManager& InputManager::GetInstance()
     return instance;
 }
 
-KeyAction InputManager::GetKeyAction()
+KeyAction InputManager::GetKeyAction(GameState currentState)
 {
     int keyInput = _getch();
 
-    // 일반 키 입력 처리 (임의 지정 - CASE문 변경 가능)
-    switch (keyInput)
+    // 상황에 따른 입력 분기
+    switch (currentState)
     {
-    case 'e': case 'E': // E 입력
-        return KeyAction::GAIN_EXP;
-    case 'q': case 'Q': // Q 입력
-        return KeyAction::QUIT;
-    case 's': case 'S': // S 입력
-        return KeyAction::SHOW_CARDS;
+        // 필드(FIELD) 상태 - 맵 이동 중 키 입력 처리
+    case GameState::FIELD:
+    {
+        switch (keyInput)
+        {
+        case 'e': case 'E': return KeyAction::GAIN_EXP;
+        case 'q': case 'Q': return KeyAction::QUIT;
+        case 's': case 'S': return KeyAction::SHOW_CARDS;
+        }
+        if (keyInput == 224)
+        { // 방향키
+            keyInput = _getch();
+            switch (keyInput)
+            {
+            case 72: return KeyAction::MOVE_UP;
+            case 80: return KeyAction::MOVE_DOWN;
+            case 75: return KeyAction::MOVE_LEFT;
+            case 77: return KeyAction::MOVE_RIGHT;
+            }
+        }
+        break; // FIELD 상태 처리 끝
     }
 
-    // 방향키 입력 처리 (224로 시작하는 특수 키)
-    if (keyInput == 224) {
-        keyInput = _getch(); // 224 이후 방향키 값
-        switch (keyInput) {
-        case 72: return KeyAction::MOVE_UP;
-        case 80: return KeyAction::MOVE_DOWN;
-        case 75: return KeyAction::MOVE_LEFT;
-        case 77: return KeyAction::MOVE_RIGHT;
+    // 상점(MERCHANT) 상태 - 맵 이동 중 키 입력 처리
+    case GameState::MERCHANT:
+    {
+        switch (keyInput)
+        { // 상점에 필요한 기능 추가해주세요.
+          // 카드 구매, 판매 등등
+        case 'q': case 'Q': return KeyAction::QUIT; // 상점 나가기
         }
+        if (keyInput == 224) { // 방향키로 아이템 선택
+            keyInput = _getch();
+            switch (keyInput)
+            {
+                // "InputManager.h"의 "enum class KeyAction" 필요한 액션 추가 후 사용
+                // case 72: return KeyAction::NEXT_ITEM; // 위 아이템 (에시)
+                // case 80: return KeyAction::NEXT_ITEM; // 아래 아이템 (에시)
+            }
+        }
+        break; // MERCHANT 상태 처리 끝
+    }
+
+    // 전투(BATTLE) 상태 - 맵 이동 중 키 입력 처리
+    case GameState::BATTLE:
+    {
+        switch (keyInput)
+        { // 전투에 필요한 기능 추가해주세요.
+        // "InputManager.h"의 "enum class KeyAction" 필요한 액션 추가 후 사용
+        // case 'a': case 'A': return KeyAction::ATTACK; // (예시)
+        // case 'd': case 'D': return KeyAction::DEFEND; // (예시)
+        // case 's': case 'S': return KeyAction::SHOW_CARDS; // (예시)
+        }
+        break; // BATTLE 상태 처리 끝
+    }
     }
 
     return KeyAction::INVALID; // 유효 키 외 입력 방지
